Compile for the C11 dialect of C
Add `-std=gnu11' (C11 + GNU extensions) to the Makefile. This is already
the default on gcc 5.4.0, as shipped with Ubuntu 16.04. Raspbian Jessie,
however, still ships gcc 4.9.2, which compiles by default with
`-std=gnu90'.

Either C99 or C11 is needed for variable declarations to be valid in the
initialization of `for' loops.