Evaluate sequences of expressions stored in files
The expression plugin can be provided with the name of a file containing
a conversion law in the form of a sequence of expressions. Each line in
the file should be of the form:

    variable_name = expression

save for the last one that should only provide the expression for the
temperature. See doc/protocol.html for details.