Do not drop root privileges
Root privileges will be needed when the client issues a `start' or
`Start' command, so they should not be relinquished on startup.

The code for dropping privileges had no effect if running from sudo, but
it made the program fail if the executable was suid root and was started
with a non privileged real user id.