BUGFIX: Early termination without libreadline
If compiled without libreadline (WITH_READLINE not set in the Makefile)
and run in shell mode, the program would display "quit" and terminate as
soon as it processed the "start" command. This is due to fgets()
returning NULL when it gets interrupted, which our naive readline()
implementation interpreted as an end-of-file condition.

We now repeat the read as long as errno is EINTR.