Avoid gcc warning -Wstringop-overflow
On line 446:

    channel_extras->conversion[0] = '\0';

GCC 10 warns:

    writing 1 byte into a region of size 0 [-Wstringop-overflow=]

It seems that GCC did not notice that, at this point in the code, the
number of parameters in the command (cmd->n_param) must be larger than
one. It thus cannot prove that the buffer channel_extras->conversion was
allocated a non-zero size.

Prevent the warning by hinting GCC that, at this point, cmd->n_param
must be larger than one.