Add standalone program for testing plugins
Typing `make test-plugin' in the plugins directory builds a program that
can be used for testing plugins without trmc2d. It must be run in the
directory containing the plugin to be tested, as:

    ./test-plugin plugin function parameters start stop step

where `plugin' is the plugin name, without the ".so" file extension.

Also, remove the (now useless) conditionally-compiled test code at the
bottom of interpolate-linear.c.