interpolate.c: do not abort on GSL errors
By default, the GSL aborts the program if it encounters an error, under
the assumption that the user will forget to check the error codes
returned by the library. We do check the error codes, so kindly ask the
library to not abort our program.

This fixes a crash that happened when a resistance-to-temperature
conversion table had duplicate values of resistance.